requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
ultralytics
opencv-python
pillow
requests
PyPDF2
docx2txt
